% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{spinnerLasso}
\alias{spinnerLasso}
\title{Spinner Lasso only, internal}
\usage{
spinnerLasso(y, SVDAx, lambdaL, WGTs, solOptions)
}
\arguments{
\item{solOptions}{}
}
\description{
Spinner Lasso only, internal
}
