% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{spinnerNuclear}
\alias{spinnerNuclear}
\title{Spinner Nuclear Norm only, internal}
\usage{
spinnerNuclear(y, SVDAx, lambda_N, solOptions)
}
\arguments{
\item{solOptions}{}
}
\description{
Spinner Nuclear Norm only, internal
}
