% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{working_example}
\alias{working_example}
\title{Example data for spinner regression}
\format{
\subsection{\code{working_example}}{

A list with two subobjects:
\describe{
\item{AA}{A 40 x 40 x 100 array containing 100 matrices for regression}
\item{y}{A length 100 vector giving the response values}
...
}
}
}
\source{
generated, bespoke
}
\usage{
working_example
}
\description{
A synthetic dataset with
}
\keyword{datasets}
